refactor: Move the mesh stuff out of numerics and into its own directory
