+++ examples/simple_conduction/with_input_file.cpp
@@ -15,7 +15,7 @@
 #include "serac/physics/state/state_manager.hpp"
 #include "serac/infrastructure/initialize.hpp"
 #include "serac/infrastructure/terminator.hpp"
-#include "serac/numerics/mesh_utils.hpp"
+#include "serac/mesh/mesh_utils.hpp"
 #include "serac/serac_config.hpp" // for SERAC_REPO_DIR
 
 const auto input_file = SERAC_REPO_DIR "/examples/simple_conduction/conduction.lua";
