+++ src/drivers/serac.cpp
@@ -19,14 +19,14 @@
 
 #include "axom/core.hpp"
 #include "mfem.hpp"
-#include "serac/coefficients/loading_functions.hpp"
+#include "serac/physics/coefficients/loading_functions.hpp"
 #include "serac/infrastructure/cli.hpp"
 #include "serac/infrastructure/initialize.hpp"
 #include "serac/infrastructure/input.hpp"
 #include "serac/infrastructure/logger.hpp"
 #include "serac/infrastructure/output.hpp"
 #include "serac/infrastructure/terminator.hpp"
-#include "serac/numerics/mesh_utils.hpp"
+#include "serac/mesh/mesh_utils.hpp"
 #include "serac/physics/thermal_solid.hpp"
 #include "serac/numerics/equation_solver.hpp"
 #include "serac/physics/state/state_manager.hpp"
