+++ src/serac/numerics/functional/tests/functional_comparisons.cpp
@@ -12,8 +12,8 @@
 #include "axom/slic/core/SimpleLogger.hpp"
 #include "serac/infrastructure/input.hpp"
 #include "serac/serac_config.hpp"
+#include "serac/mesh/mesh_utils_base.hpp"
 #include "serac/numerics/expr_template_ops.hpp"
-#include "serac/numerics/mesh_utils_base.hpp"
 #include "serac/numerics/stdfunction_operator.hpp"
 #include "serac/numerics/functional/functional.hpp"
 #include "serac/numerics/functional/tensor.hpp"
