+++ src/serac/physics/coefficients/sensitivity_coefficients.cpp
@@ -4,7 +4,7 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
-#include "serac/coefficients/sensitivity_coefficients.hpp"
+#include "serac/physics/coefficients/sensitivity_coefficients.hpp"
 #include "serac/physics/materials/solid_utils.hpp"
 
 namespace serac::mfem_ext {
