+++ src/serac/physics/solid.cpp
@@ -12,8 +12,8 @@
 #include "serac/physics/integrators/wrapper_integrator.hpp"
 #include "serac/physics/state/state_manager.hpp"
 #include "serac/numerics/expr_template_ops.hpp"
-#include "serac/numerics/mesh_utils.hpp"
-#include "serac/coefficients/sensitivity_coefficients.hpp"
+#include "serac/mesh/mesh_utils.hpp"
+#include "serac/physics/coefficients/sensitivity_coefficients.hpp"
 
 namespace serac {
