+++ src/serac/physics/tests/serac_dtor.cpp
@@ -14,7 +14,7 @@
 #include <gtest/gtest.h>
 #include "mfem.hpp"
 
-#include "serac/numerics/mesh_utils.hpp"
+#include "serac/mesh/mesh_utils.hpp"
 #include "serac/physics/state/state_manager.hpp"
 #include "serac/serac_config.hpp"
