+++ src/serac/physics/tests/test_utilities.cpp
@@ -11,7 +11,7 @@
 #include "axom/core.hpp"
 
 #include "serac/infrastructure/input.hpp"
-#include "serac/numerics/mesh_utils.hpp"
+#include "serac/mesh/mesh_utils.hpp"
 #include "serac/physics/solid.hpp"
 #include "serac/physics/thermal_conduction.hpp"
 #include "serac/physics/thermal_solid.hpp"
